[#529] Replace OSLogger_ with OSLogger from the SDK
Closes #529
