+++ secant/SecantApp.swift
@@ -20,7 +20,7 @@ final class AppDelegate: NSObject, UIApplicationDelegate {
         _ application: UIApplication,
         didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]? = nil
     ) -> Bool {
-        walletLogger = OSLogger_(logLevel: .debug, category: LoggerConstants.walletLogs)
+        walletLogger = OSLogger(logLevel: .debug, category: LoggerConstants.walletLogs)
         // set the default behavior for the NSDecimalNumber
         NSDecimalNumber.defaultBehavior = Zatoshi.decimalHandler
         rootViewStore.send(.initialization(.appDelegate(.didFinishLaunching)))
