+++ secant/Utils/Logging/TCALogging.swift
@@ -1,5 +1,5 @@
 //
-//  TCALogger.swift
+//  TCALogging.swift
 //  secant-testnet
 //
 //  Created by Lukáš Korba on 23.01.2023.
@@ -7,11 +7,10 @@
 
 import Foundation
 import os
+import ZcashLightClientKit
 
-class TCALogger: OSLogger_ { }
-
-extension TCALogger {
-    static let live = TCALogger(logLevel: .debug, category: LoggerConstants.tcaLogs)
+extension OSLogger {
+    static let live = OSLogger(logLevel: .debug, category: LoggerConstants.tcaLogs)
 
     func tcaDebug(_ message: String) {
         guard let oslog else { return }
